limpiando codigo de materias
